Tooling: Schritt 0c.3 + 0c.4 (host-only strict warnings + ASan/UBSan)
Schließt die zwei aus Group A nach Schritt 11 verschobenen Sub-Schritte
und synchronisiert die Plan-Doku.

env:native baut jetzt unsere src/+test/-TUs mit -Werror, -Wpedantic,
-Wshadow, -Wcast-align/-qual, -Wmisleading-indentation, -Wnull-dereference,
-Wdouble-promotion, -Wformat=2, -Wundef, -Winit-self,
-Wzero-as-null-pointer-constant — plus ASan + UBSan auf der Link-Linie.
Library-Header (ArduinoJson, Unity) sind via -isystem ausgeklammert.

Scope-Reduktion gegenüber Plan-Sketch §0c.3 (alle Memory-konform, keine
SCons-Hooks für selektives Flag-Routing, siehe AwesomeStudioPedal-Vergleich):
- C++-only Flags raus (PIO generiert unity_config.c → cc1 errors auf
  C-Files). Coverage via clang-tidy.
- GCC-only Flags raus (make tidy reuses Flags via compile_commands.json,
  clang-tidy rejected unknown options). Coverage via bugprone-branch-clone.
- -Wconversion/-Wsign-conversion raus (unity_config.c:39 trippt putchar(c)).
  Coverage via ASan/UBSan + .clang-tidy.

Drive-by: DEFAULT_FILTER_HEALTH_DEAD_AFTER + CycleConfig-Feld auf uint8_t
(Konsistenz mit FilterHealth(uint8_t)-Ctor, vom strict-Build gefangen).

make ci grün (format-check + lint + tidy + test-native mit Sanitizern +
ESP32-build); 138/138 native Tests grün.

+++ src/logic/cycle_runner.h
@@ -26,7 +26,7 @@ constexpr unsigned DEFAULT_COLD_BOOT_GIVEUP_SLEEP_S = 300;
 constexpr unsigned DEFAULT_POLL_INTERVAL_S = 30;
 constexpr int DEFAULT_STALE_THRESHOLD_S = 180;
 constexpr int DEFAULT_NTP_INTERVAL_S = 86400;
-constexpr int DEFAULT_FILTER_HEALTH_DEAD_AFTER = 3;
+constexpr uint8_t DEFAULT_FILTER_HEALTH_DEAD_AFTER = 3;
 constexpr unsigned DEFAULT_LONG_SLEEP_FOR_NIGHTLY_CLEAN_S = 4U * 3600U;
 constexpr int DEFAULT_NIGHTLY_DEEP_CLEAN_INTERVAL_S = 20 * 3600;
 constexpr unsigned DEFAULT_BTN_LONG_PRESS_MS = 2000;
@@ -44,7 +44,7 @@ struct CycleConfig {
   unsigned poll_interval_s = DEFAULT_POLL_INTERVAL_S;
   int stale_threshold_s = DEFAULT_STALE_THRESHOLD_S;
   int ntp_interval_s = DEFAULT_NTP_INTERVAL_S;
-  int filter_health_dead_after = DEFAULT_FILTER_HEALTH_DEAD_AFTER;
+  uint8_t filter_health_dead_after = DEFAULT_FILTER_HEALTH_DEAD_AFTER;
   int wake_before_bus_s = DEFAULT_WAKE_BEFORE_BUS_S;
   int boot_margin_s = DEFAULT_BOOT_MARGIN_S;
   int active_threshold_s = DEFAULT_ACTIVE_THRESHOLD_S;
